+++ src/lib.rs
@@ -1,0 +1,48 @@
+#[macro_use]
+extern crate minimad;
+#[macro_use]
+extern crate lazy_static;
+#[macro_use]
+extern crate log;
+#[macro_use]
+extern crate lazy_regex;
+
+pub mod app;
+pub mod app_context;
+pub mod app_state;
+pub mod browser_states;
+pub mod browser_verbs;
+pub mod cli;
+pub mod command_parsing;
+pub mod commands;
+pub mod conf;
+pub mod displayable_tree;
+pub mod external;
+pub mod file_sizes;
+pub mod flat_tree;
+pub mod git_ignore;
+pub mod help_content;
+pub mod help_states;
+pub mod help_verbs;
+pub mod io;
+pub mod mad_skin;
+pub mod matched_string;
+pub mod permissions;
+pub mod screens;
+pub mod shell_bash;
+pub mod shell_fish;
+pub mod shell_install;
+pub mod skin;
+pub mod skin_conf;
+pub mod status;
+pub mod task_sync;
+pub mod tree_build;
+pub mod tree_options;
+pub mod verb_conf;
+pub mod verb_invocation;
+pub mod verb_store;
+pub mod verbs;
+pub mod errors;
+pub mod fuzzy_patterns;
+pub mod patterns;
+pub mod regex_patterns;